refactor: modularize app.js
