+++ listeners/actions/index.js
@@ -1,0 +1,5 @@
+const { handle_feedback } = require('./feedback.js');
+
+export const register = (app) => {
+  app.action('feedback', handle_feedback);
+};
