+++ listeners/events/index.js
@@ -1,0 +1,5 @@
+const { handleAppMention } = require('./app_mentioned.js');
+
+export const register = (app) => {
+  app.event('app_mention', handleAppMention);
+};
